refactored MessagesTransformer and subclasses location
