+++ src/quickapp/common/__init__.py
@@ -1,5 +1,6 @@
 from ._di_types import DIAL_API_KEY, DIAL_BEARER, RESPONSE_FORMAT
 from .base_initializer import BaseInitializer, InitializerType
+from .base_transformer import MessagesTransformer
 from .completion_result import CompletionResult
 from .staged_base_tool import StagedBaseTool
 from .timed_stage_wrapper import TimedStageWrapper
