+++ src/quickapp/internal_tooling/attachment_notification_tooling/__init__.py
@@ -1,0 +1,5 @@
+from quickapp.internal_tooling.attachment_notification_tooling._attachment_notification_injector import (
+    AttachmentNotificationInjector,
+)
+
+__all__ = ["AttachmentNotificationInjector"]
