+++ src/tests/unit_tests/agent_tests/test_attachment_notification_injector.py
@@ -3,7 +3,7 @@ import json
 from aidial_sdk.chat_completion import Attachment, CustomContent, Message, Role
 from pydantic.v1 import StrictStr
 
-from quickapp.agent.processors.pre_transformers import AttachmentNotificationInjector
+from quickapp.internal_tooling.attachment_notification_tooling import AttachmentNotificationInjector
 from quickapp.config.context import FileContextConfig
 from quickapp.internal_tooling.attachment_notification_tooling._tool_configs import (
     AVAILABLE_CONTEXT_TOOL_NAME,
